+++ aws_emr_launch/lambda_sources/emr_utilities/load_cluster_configuration/lambda_source.py
@@ -107,7 +107,8 @@ def handler(event, context):
         cluster_configuration['JobFlowRole'] = emr_profile['Roles']['InstanceRole'].split('/')[-1]
         cluster_configuration['ServiceRole'] = emr_profile['Roles']['ServiceRole'].split('/')[-1]
         cluster_configuration['AutoScalingRole'] = emr_profile['Roles']['AutoScalingRole'].split('/')[-1] \
-            if len(cluster_configuration['Instances'].get('InstanceGroups', [])) > 0 else None
+            if cluster_configuration['Instances'].get('InstanceGroups', []) \
+            and len(cluster_configuration['Instances'].get('InstanceGroups', [])) > 0 else None
         cluster_configuration['Tags'] = tags
         cluster_configuration['Instances']['EmrManagedMasterSecurityGroup'] = \
             emr_profile['SecurityGroups']['MasterGroup']